+++ banyand/stream/benchmark_test.go
@@ -344,6 +344,7 @@ func generateStreamQueryOptions(p parameter, index mockIndex) pbv1.StreamQueryOp
 
 func BenchmarkFilter(b *testing.B) {
 	b.ReportAllocs()
+	ctx := context.TODO()
 	for _, p := range pList {
 		esList, docsList, idx := generateData(p)
 		db := write(b, p, esList, docsList)
@@ -351,24 +352,25 @@ func BenchmarkFilter(b *testing.B) {
 		sqo := generateStreamQueryOptions(p, idx)
 		sqo.Order = nil
 		b.Run("filter-"+p.scenario, func(b *testing.B) {
-			res, err := s.Query(context.TODO(), sqo)
+			res, err := s.Query(ctx, sqo)
 			require.NoError(b, err)
-			logicalstream.BuildElementsFromStreamResult(res)
+			logicalstream.BuildElementsFromStreamResult(ctx, res)
 		})
 	}
 }
 
 func BenchmarkSort(b *testing.B) {
 	b.ReportAllocs()
+	ctx := context.TODO()
 	for _, p := range pList {
 		esList, docsList, idx := generateData(p)
 		db := write(b, p, esList, docsList)
 		s := generateStream(db)
 		sqo := generateStreamQueryOptions(p, idx)
 		b.Run("sort-"+p.scenario, func(b *testing.B) {
-			res, err := s.Query(context.TODO(), sqo)
+			res, err := s.Query(ctx, sqo)
 			require.NoError(b, err)
-			logicalstream.BuildElementsFromStreamResult(res)
+			logicalstream.BuildElementsFromStreamResult(ctx, res)
 		})
 	}
 }